requirements.txt
# MoLAQ 依赖
# 核心依赖已通过 ~/vllm 虚拟环境提供，此处仅列出额外需要确认的包
torch>=2.1.0
transformers>=4.45.0
pillow
tqdm
pytest
# llm-compressor 和 vllm 通过 ~/vllm 虚拟环境提供，无需单独安装
